add cjs, rename dcx folder to core, add cjs integration test
