+++ src/core/agent.ts
