+++ src/core/did-dht-manager.ts
