+++ src/core/identity-vault.ts
