+++ src/core/manager.ts
