--- src/dcx/index.ts
@@ -1,6 +1,0 @@
-export * from './manager.js';
-export * from './config.js';
-export * from './server.js';
-export * from './handlers.js';
-import server from './server.js';
-export default server;
