+++ src/index.ts
@@ -1,6 +1,11 @@
-export * from './dcx/index.js';
+export * from './core/index.js';
 export * from './protocol/index.js';
+export * from './schemas/index.js';
 export type * from './types/index.js';
 export * from './utils/index.js';
-import server from './dcx/index.js';
-export default server;
+
+import DcxServer from './core/index.js';
+export default DcxServer;
+
+import { server } from './core/index.js';
+export { server };
