+++ src/utils/index.ts
@@ -2,7 +2,10 @@ export * from './cipher.js';
 export * from './dwn.js';
 export * from './error.js';
 export * from './file-system.js';
+export * from './identity-vault.js';
+
 export * from './json.js';
 export * from './logger.js';
 export * from './objects.js';
+export * from './string.js';
 export * from './time.js';
