+++ tests/server.integration.ts
@@ -1,4 +1,4 @@
-import DcxServer from '../src/dcx/index.js';
+import DcxServer from '../src/index.js';
 import ExampleManifest from '../EXAMPLE-MANIFEST.json';
 
 const server = new DcxServer();
@@ -6,9 +6,9 @@ const server = new DcxServer();
 server.use('manifest', ExampleManifest);
 server.use('provider',
   {
-    id: ExampleManifest.output_descriptors[0].id,
-    method: 'POST',
-    endpoint: 'http://localhost:4000/api/v1/vc/data'
+    id       : ExampleManifest.output_descriptors[0].id,
+    method   : 'POST',
+    endpoint : 'http://localhost:4000/api/v1/vc/data'
   }
 );
 server.use('dwn', 'http://localhost:3000/');
